Remove web socket changes
