+++ packages/wxt/src/virtual/reload-html.ts
@@ -1,12 +1,16 @@
 import { logger } from '../sandbox/utils/logger';
-import { getDevServerWebSocket } from '../sandbox/dev-server-websocket';
+import { setupWebSocket } from './utils/setup-web-socket';
 
 if (import.meta.env.COMMAND === 'serve') {
   try {
-    const ws = getDevServerWebSocket();
-    ws.addWxtEventListener('wxt:reload-page', (event) => {
-      // "popup.html" === "/popup.html".substring(1)
-      if (event.detail === location.pathname.substring(1)) location.reload();
+    setupWebSocket((message) => {
+      if (message.event === 'wxt:reload-page') {
+        // We need to remove the initial slash from the path to compare correctly
+        // "popup.html" === "/popup.html".substring(1)
+        if (message.data === location.pathname.substring(1)) {
+          location.reload();
+        }
+      }
     });
   } catch (err) {
     logger.error('Failed to setup web socket connection with dev server', err);
