+++ packages/wxt/src/virtual/utils/reload-content-scripts.ts
@@ -1,7 +1,16 @@
 import { browser } from 'wxt/browser';
 import { logger } from '../../sandbox/utils/logger';
 import { MatchPattern } from 'wxt/sandbox';
-import type { ReloadContentScriptPayload } from '../../sandbox/dev-server-websocket';
+
+interface ContentScript {
+  matches: string[];
+  js?: string[];
+  css?: string[];
+}
+interface ReloadContentScriptPayload {
+  registration?: 'manifest' | 'runtime';
+  contentScript: ContentScript;
+}
 
 export function reloadContentScript(payload: ReloadContentScriptPayload) {
   const manifest = browser.runtime.getManifest();
@@ -22,8 +31,6 @@ export async function reloadContentScriptMv3({
     await reloadManifestContentScriptMv3(contentScript);
   }
 }
-
-type ContentScript = ReloadContentScriptPayload['contentScript'];
 
 export async function reloadManifestContentScriptMv3(
   contentScript: ContentScript,
